Moved the code from the viz module into the stix2viz package. Removed the viz module.  Also replaced occurrences of "stix2jupyter" with "stix2viz" in application.js.
